+++ src/orbitops/cli.py
@@ -1,8 +1,9 @@
 import sys
-
-from . import api, etc, propagation
+from datetime import UTC, datetime, timedelta
 
 from rich.console import Console
+
+from . import api, etc, propagation
 
 console = Console()
 
@@ -16,6 +17,7 @@ VALID_COMMANDS = (
     "search",
     "distance",
     "watch",
+    "gtrack",
 )
 
 
@@ -38,7 +40,9 @@ def main() -> None:
         # Invalid command
         if command not in VALID_COMMANDS:
             console.print(f"[bold red]Invalid command: {command}[/bold red]")
-            console.print("[dim yellow]Use 'orbitops help' to view available commands.[/dim yellow]")
+            console.print(
+                "[dim yellow]Use 'orbitops help' to view available commands.[/dim yellow]"
+            )
             return
 
         # Search requires a name
@@ -49,23 +53,30 @@ def main() -> None:
 
         # Distance requires two catalog numbers
         if command == "distance" and len(sys.argv) < 4:
-            console.print("[bold red]Two satellite catalog numbers are required.[/bold red]")
-            console.print("[dim yellow]Usage: orbitops distance <CATNR1> <CATNR2>[/dim yellow]")
+            console.print(
+                "[bold red]Two satellite catalog numbers are required.[/bold red]"
+            )
+            console.print(
+                "[dim yellow]Usage: orbitops distance <CATNR1> <CATNR2>[/dim yellow]"
+            )
             return
 
         # All remaining commands require one catalog number
         if command not in ("search", "distance"):
-
             if len(sys.argv) < 3:
                 console.print("[bold red]Missing satellite catalog number.[/bold red]")
-                console.print(f"[dim yellow]Usage: orbitops {command} <CATNR>[/dim yellow]")
+                console.print(
+                    f"[dim yellow]Usage: orbitops {command} <CATNR>[/dim yellow]"
+                )
                 return
 
             try:
                 catalog_number = int(sys.argv[2])
 
             except ValueError:
-                console.print("[bold red]Satellite catalog number must be an integer.[/bold red]")
+                console.print(
+                    "[bold red]Satellite catalog number must be an integer.[/bold red]"
+                )
                 return
 
             sat_data = api.get_sat_info_tle(catalog_number)
@@ -80,11 +91,7 @@ def main() -> None:
             sat_name, tle_line1, tle_line2 = sat_data
 
         if command == "teme":
-
-            position, velocity = propagation.get_teme_cartesian(
-                tle_line1,
-                tle_line2
-            )
+            position, velocity = propagation.get_teme_cartesian(tle_line1, tle_line2)
 
             print(f"\nTEME Cartesian State of {sat_name}")
             print("--------------------")
@@ -100,15 +107,12 @@ def main() -> None:
             print(f"  Z: {velocity[2]:10.3f} km/s")
 
         if command == "position":
-
             print(f"\nGeographic position of {sat_name}")
             print("--------------------")
 
-            latitude, longitude, altitude = (
-                propagation.get_geographic_position(
-                    tle_line1,
-                    tle_line2,
-                )
+            latitude, longitude, altitude = propagation.get_geographic_position(
+                tle_line1,
+                tle_line2,
             )
 
             print(f"Latitude:  {latitude:.4f}°")
@@ -116,7 +120,6 @@ def main() -> None:
             print(f"Altitude:  {altitude:.2f} km")
 
         if command == "info":
-
             data = api.get_satcat_data(catalog_number)
 
             if not data:
@@ -130,7 +133,6 @@ def main() -> None:
                 print(f"{key}: {value}")
 
         if command == "search":
-
             results = api.search_by_name(sys.argv[2])
 
             if not results:
@@ -146,26 +148,45 @@ def main() -> None:
                 print(f"{key}: {value}")
 
         if command == "distance":
-
             try:
                 catalog_number_1 = int(sys.argv[2])
                 catalog_number_2 = int(sys.argv[3])
 
             except ValueError:
-                console.print("[bold red]Satellite catalog numbers must be integers.[/bold red]")
+                console.print(
+                    "[bold red]Satellite catalog numbers must be integers.[/bold red]"
+                )
                 return
 
-            api.get_distance_sats(
-                catalog_number_1,
-                catalog_number_2
-            )
+            api.get_distance_sats(catalog_number_1, catalog_number_2)
 
         if command == "watch":
-
             api.watch(catalog_number)
 
+        if command == "gtrack":
+            minutes = int(sys.argv[3])
+
+            start_time = datetime.now(UTC)
+
+            times = []
+
+            for minute in range(minutes + 1):
+                times.append(start_time + timedelta(minutes=minute))
+
+            ground_track = etc.generate_ground_track(tle_line1, tle_line2, times)
+
+            etc.plot_ground_track(
+                ground_track,
+                sat_name,
+                minutes,
+                tle_line1,
+                tle_line2,
+            )
+
     except IndexError as error:
-        console.print(f"[bold red]Invalid satellite catalog number (CATNR): {error}[/bold red]")
+        console.print(
+            f"[bold red]Invalid satellite catalog number (CATNR): {error}[/bold red]"
+        )
 
     except KeyboardInterrupt:
         console.print("[dim yellow]\nOrbitOps stopped.[/dim yellow]")